+++ app/app/(root)/page.tsx
@@ -1,19 +1,18 @@
-import { AuthBtn } from "@/components/btns/auth-btn"
-import { authOptions } from "@/lib/auth/auth-options"
-import { getServerSession } from "next-auth"
-
+import { auth } from "@/auth";
+import { AuthBtn } from "@/components/btns/auth-btn";
 
 export default async function Page() {
-  const session = await getServerSession(authOptions)
-  if (!session) return (
-    <div className="h-screen flex items-center">
-      <div className="container mx-auto h-3/4">
-        <div className="flex justify-center">
-        <AuthBtn callbackUrl="/" />
+  const session = await auth();
+  if (!session)
+    return (
+      <div className="h-screen flex items-center">
+        <div className="container mx-auto h-3/4">
+          <div className="flex justify-center">
+            <AuthBtn callbackUrl="/" />
+          </div>
         </div>
       </div>
-    </div>
-  )
+    );
   return (
     <div className="h-screen flex items-center">
       <div className="container mx-auto h-3/4">
@@ -22,5 +21,5 @@ export default async function Page() {
         </p>
       </div>
     </div>
-  )
+  );
 }
