+++ app/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,1 @@
-import NextAuth from "next-auth"
-import { authOptions } from "@/lib/auth/auth-options"
-
-const handler = NextAuth(authOptions)
-
-export { handler as GET, handler as POST }
+export { GET, POST } from "@/auth";
