+++ app/components/btns/auth-btn/index.tsx
@@ -1,15 +1,12 @@
-import { authOptions } from "@/lib/auth/auth-options"
-import { getServerSession } from "next-auth"
-import { SignOutBtn } from "../sign-out-btn"
-import { SignInBtn } from "../sign-in-btn"
+import { SignOutBtn } from "../sign-out-btn";
+import { SignInBtn } from "../sign-in-btn";
+import { auth } from "@/auth";
 
-export const AuthBtn = async ({ callbackUrl }: { callbackUrl: string}) => {
-  const session = await getServerSession(authOptions)
+export const AuthBtn = async ({ callbackUrl }: { callbackUrl: string }) => {
+  const session = await auth();
   return (
     <div>
-      {session 
-        ? <SignOutBtn />
-        : <SignInBtn callbackUrl={callbackUrl} />}
+      {session ? <SignOutBtn /> : <SignInBtn callbackUrl={callbackUrl} />}
     </div>
-  )
-}
+  );
+};
