+++ app/types/next-auth.d.ts
@@ -3,17 +3,18 @@
  * It defines additional properties for the `Session` and `JWT` objects, and
  * extends the `Profile` interface to include properties specific to the 42 School provider.
  * See: https://next-auth.js.org/getting-started/typescript
+ * Note: Docs not updated on how to augment JWT
  */
-import NextAuth, { DefaultSession, Profile } from "next-auth"
+import NextAuth, { DefaultSession, Profile } from "next-auth";
 import { FortyTwoProfile } from "next-auth/providers/42-school";
-import { JWT } from "next-auth/jwt"
+import { JWT } from "next-auth/jwt";
 
 declare module "next-auth" {
   interface Session {
     user: {
-      login: string
-      isStaff: boolean
-    } & DefaultSession["user"]
+      login: string;
+      isStaff: boolean;
+    } & DefaultSession["user"];
   }
   interface Profile extends FortyTwoProfile {}
 }
@@ -21,8 +22,8 @@ declare module "next-auth" {
 declare module "next-auth/jwt" {
   interface JWT {
     user: {
-      login: string
-      isStaff: boolean
-    }
+      login: string;
+      isStaff: boolean;
+    };
   }
 }
